refactor: update imports to use UserType from entities and clean up unused imports
