+++ src/main/kotlin/data/mapper/UserMapper.kt
@@ -2,7 +2,7 @@ package org.baghdad.data.mapper
 
 import org.baghdad.data.dto.UserDto
 import org.baghdad.logic.model.entities.UserEntity
-import org.baghdad.logic.model.enums.UserType
+import org.baghdad.logic.model.entities.UserType
 
 fun UserDto.toDomain(): UserEntity {
     return UserEntity(
